requirements.txt
numpy
scikit-learn
tensorflow
matplotlib
seaborn
tqdm
joblib
pillow
opencv-python
pytest
